wta mingration functionality

+++ dto/src/main/java/com/kairos/response/dto/web/UnitPositionDTO.java
@@ -4,7 +4,6 @@ import com.kairos.persistence.model.enums.EmploymentCategory;
 import org.hibernate.validator.constraints.Range;
 
 import javax.validation.constraints.NotNull;
-import java.math.BigInteger;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -38,7 +37,7 @@ public class UnitPositionDTO {
     @NotNull(message = "employmentTypeCategory can't be null")
     private EmploymentCategory employmentTypeCategory;
     @NotNull(message = "wta can't be null")
-    private BigInteger wtaId;
+    private Long wtaId;
     private Long ctaId;
     @NotNull(message = "staffId is missing")
     @Range(min = 0, message = "staffId is missing")
@@ -84,7 +83,7 @@ public class UnitPositionDTO {
 
 
     public UnitPositionDTO(Long positionCodeId, Long expertiseId, Long startDateMillis, Long endDateMillis, int totalWeeklyHours, Long employmentTypeId,
-                           Long staffId, BigInteger wtaId, Long ctaId, Long unitId, Long timeCareExternalId) {
+                           Long staffId, Long wtaId, Long ctaId, Long unitId, Long timeCareExternalId) {
         this.positionCodeId = positionCodeId;
         this.expertiseId = expertiseId;
         this.employmentTypeId = employmentTypeId;
@@ -184,11 +183,11 @@ public class UnitPositionDTO {
         this.hourlyWages = hourlyWages;
     }
 
-    public BigInteger getWtaId() {
+    public Long getWtaId() {
         return wtaId;
     }
 
-    public void setWtaId(BigInteger wtaId) {
+    public void setWtaId(Long wtaId) {
         this.wtaId = wtaId;
     }
